requirements.txt
pika==1.3.1
python-dotenv
